cockroach no observes the mouse cursor

+++ src/ey/mrndesign/matned/client/model/object/GameObject.java
@@ -3,8 +3,8 @@ package ey.mrndesign.matned.client.model.object;
 public interface GameObject {
 
 
-    int getSizeX();
-    int getSizeY();
-    int getPositionX();
-    int getPositionY();
+    double getSizeX();
+    double getSizeY();
+    double getPositionX();
+    double getPositionY();
 }
