+++ src/ey/mrndesign/matned/client/model/object/MovingObject.java
@@ -2,7 +2,7 @@ package ey.mrndesign.matned.client.model.object;
 
 public interface MovingObject extends GameObject {
 
-    void setPositionX(int positionX);
-    void setPositionY(int positionY);
+    void setPositionX(double positionX);
+    void setPositionY(double positionY);
 
 }
